+++ mandatory/lexer/lexical_analyzer.c
@@ -1,58 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexical.c                                          :+:      :+:    :+:   */
+/*   lexical_analyzer.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: gacalaza <gacalaza@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 17:04:43 by gacalaza          #+#    #+#             */
-/*   Updated: 2023/11/27 20:45:56 by gacalaza         ###   ########.fr       */
+/*   Updated: 2023/12/01 14:33:24 by gacalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-
-int	lexer(t_data *data)
-{
-	if (data->tokens->type == 3)
-		printf("heya\n");
-	return (1);
-}
-
-int	is_a_quote(int type)
-{
-	if (type == QUOTE_DOUBLE)
-		return (QUOTE_DOUBLE);
-	if (type == QUOTE_SINGLE)
-		return (QUOTE_SINGLE);
-	return (FALSE);
-}
-
-int	has_quotes(t_data *data)
+int	has_pipe(t_token *tokens)
 {
 	t_token	*tmp;
 	int		check;
 
-	tmp = data->tokens;
-	check = 0;
-	while (tmp)
-	{
-		if (is_a_quote(tmp->type))
-			check = has_another_quote(tmp, tmp->type);
-		if (check > 0)
-			return (TRUE);
-		tmp = tmp->next;
-	}
-	return (FALSE);
-}
-
-int	has_pipe(t_data *data)
-{
-	t_token	*tmp;
-	int		check;
-
-	tmp = data->tokens;
+	tmp = tokens;
 	check = 0;
 	while (tmp)
 	{
@@ -61,4 +26,54 @@ int	has_pipe(t_data *data)
 		tmp = tmp->next;
 	}
 	return (check);
+}
+
+int	has_rdrct_pipe(t_token *tokens)
+{
+	t_token	*tmp;
+	int		redirects;
+
+	tmp = tokens;
+	redirects = 0;
+	while (tmp)
+	{
+		if (tmp->type == REDIRECT_IN || tmp->type == REDIRECT_OUT)
+			redirects++;
+		if (tmp->type == PIPE)
+			break ;
+		tmp = tmp->next;
+	}
+	return (redirects);
+}
+
+int	has_redirect(t_token *tokens)
+{
+	t_token	*temp;
+	int		redirects;
+
+	temp = tokens;
+	redirects = 0;
+	while (temp)
+	{
+		if (temp->type == 1 || temp->type == 2)
+			redirects++;
+		temp = temp->next;
+	}
+	return (redirects);
+}
+
+int	has_dredirect(t_token *tokens)
+{
+	t_token	*temp;
+	int		redirects;
+
+	temp = tokens;
+	redirects = 0;
+	while (temp)
+	{
+		if (temp->type == HEREDOC || temp->type == APPEND)
+			redirects++;
+		temp = temp->next;
+	}
+	return (redirects);
 }
