+++ mandatory/lexer/lexical_five.c
@@ -1,39 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   token_utils_four.c                                 :+:      :+:    :+:   */
+/*   lexical_five.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: gacalaza <gacalaza@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/12 17:05:39 by gacalaza          #+#    #+#             */
-/*   Updated: 2023/11/10 18:25:14 by gacalaza         ###   ########.fr       */
+/*   Created: 2023/12/01 18:47:47 by gacalaza          #+#    #+#             */
+/*   Updated: 2023/12/01 18:48:53 by gacalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-int	is_special_char(char c)
+int	is_drd_case(int type)
 {
-	if (c == '*')
-		return (ASTERISK);
-	if (c == '&')
-		return (AMPERSAND);
-	if (c == '~')
-		return (TILDE);
-	if (c == '.')
-		return (DOT);
-	if (c == ';')
-		return (SEMICOLON);
-	return (0);
-}
-
-int	is_r_bracket(char c)
-{
-	if (c == '(')
-		return (R_BRACKET_O);
-	if (c == ')')
-		return (R_BRACKET_C);
-	return (0);
+	if (type == HEREDOC)
+		return (TRUE);
+	if (type == APPEND)
+		return (TRUE);
+	return (FALSE);
 }
 
 int	is_hd_c(char *str)
@@ -49,4 +34,22 @@ int	is_e_c(char *str)
 	if (find_type(&str[0]) == EXIT_STATUS)
 		return (1);
 	return (0);
+}
+
+int	is_redrt_case(int type)
+{
+	if (type == REDIRECT_IN || type == REDIRECT_OUT)
+		return (TRUE);
+	if (type == HEREDOC || type == APPEND)
+		return (TRUE);
+	return (FALSE);
+}
+
+int	is_rd_case(int type)
+{
+	if (type == REDIRECT_IN)
+		return (TRUE);
+	if (type == REDIRECT_OUT)
+		return (TRUE);
+	return (FALSE);
 }
