+++ mandatory/parse/parse_utils.c
@@ -3,63 +3,94 @@
 /*                                                        :::      ::::::::   */
 /*   parse_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ckunimur <ckunimur@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: gacalaza <gacalaza@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/03 21:14:38 by gacalaza          #+#    #+#             */
-/*   Updated: 2023/11/29 15:29:47 by ckunimur         ###   ########.fr       */
+/*   Updated: 2023/12/01 16:56:16 by gacalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-int	is_rd_case(int type)
+void	cmd_pipe(t_data *data)
 {
-	if (type == REDIRECT_IN)
-		return (TRUE);
-	if (type == REDIRECT_OUT)
-		return (TRUE);
-	return (FALSE);
-}
+	t_token	*tmp;
+	char	**all_words;
 
-int	is_drd_case(int type)
-{
-	if (type == HEREDOC)
-		return (TRUE);
-	if (type == APPEND)
-		return (TRUE);
-	return (FALSE);
-}
-
-int	is_word(int type, int check)
-{
-	if (check == 1)
+	tmp = data->tokens;
+	all_words = get_all_words(data->tokens);
+	while (tmp)
 	{
-		if (type == WORD)
-			return (TRUE);
-		if (type == EXIT_STATUS)
-			return (TRUE);
+		if (tmp->type == PIPE || tmp->next == NULL)
+		{
+			finalizepipe_cmd(data, all_words);
+			if (tmp->next != NULL)
+				all_words = fixwords(tmp->next, all_words);
+		}
+		tmp = tmp->next;
 	}
-	if (check == 2)
-	{
-		if (type != REDIRECT_IN && type != REDIRECT_OUT
-			&& type != HEREDOC && type != APPEND
-			&& type != C_SPACE)
-			return (TRUE);
-	}
-	return (FALSE);
 }
 
-void	print_array(char **array, char *type)
+char	*get_cmd(char **words)
+{
+	char	*cmd;
+
+	cmd = ft_strdup(words[0]);
+	return (cmd);
+}
+
+char	**get_args(char **words, int len)
+{
+	char	**args;
+
+	args = NULL;
+	if (len > 1)
+		return (ft_arraydup((&words[1])));
+	return (args);
+}
+
+int	count_backs(int len, char *str)
 {
 	int	i;
+	int	backs;
 
 	i = 0;
-	if (!array)
-		return ;
-	while (array[i] != NULL)
+	backs = 0;
+	while (i < len)
 	{
-		printf("%s[%i]: %s\n", type, i, array[i]);
+		if (str[i] == '\\' && str[i + 1] != '\\')
+			backs++;
+		if (i + 1 == len)
+			break ;
 		i++;
 	}
-	printf("\n");
+	return (backs);
+}
+
+char	*process_backs(char *str, int len)
+{
+	int		i;
+	int		j;
+	int		backs;
+	char	*new_str;
+
+	i = 0;
+	j = 0;
+	new_str = NULL;
+	backs = count_backs(len, str);
+	if (backs > 0)
+	{
+		new_str = malloc(sizeof (char) * (len - backs) + 1);
+		if (!new_str)
+			return_error_parse("malloc");
+		while (j < (len - backs))
+		{
+			if (str[++i] == '\\')
+				new_str[j++] = str[i];
+			else
+				new_str[j++] = str[i];
+		}
+		new_str[j] = '\0';
+	}
+	return (new_str);
 }
