+++ mandatory/utils/ft_intdup.c
@@ -1,42 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_echo.c                                          :+:      :+:    :+:   */
+/*   ft_intdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: gacalaza <gacalaza@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/17 21:04:33 by ckunimur          #+#    #+#             */
-/*   Updated: 2023/12/01 18:49:42 by gacalaza         ###   ########.fr       */
+/*   Created: 2023/12/01 11:41:14 by gacalaza          #+#    #+#             */
+/*   Updated: 2023/12/01 11:41:18 by gacalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
 
-void	ft_echo(t_data *data)
+int	*ft_intdup(int *array, int size)
 {
+	int		*array_copy;
 	int		i;
-	t_cmd	*temp;
 
-	i = 1;
-	temp = data->cmd;
-	printf("%s", temp->cmd[i]);
-	i++;
-	if (ft_array_size(data->cmd->cmd) > 1
-		&& ft_strncmp(data->cmd->cmd[1], "-n", 3) == 0)
+	if (array == NULL)
+		return (NULL);
+	i = 0;
+	array_copy = (int *)malloc(sizeof(int) * size);
+	if (array_copy == NULL)
+		return (NULL);
+	while (i < size)
 	{
-		while (temp->cmd[i])
-		{
-			printf(" %s", temp->cmd[i]);
-			i++;
-		}
+		array_copy[i] = array[i];
+		i++;
 	}
-	else
-	{
-		while (temp->cmd[i])
-		{
-			printf(" %s", temp->cmd[i]);
-			i++;
-		}
-		printf("\n");
-	}
+	return (array_copy);
 }
